+++ extensions/slack/src/search.tsx
@@ -1,11 +1,47 @@
 // This filename should be named `switch-to-channel.tsx` or something similar
 // but it's kept as `search.tsx` as changing the command's name will cause users to lose their keywords and aliases
-import { ActionPanel, Action, Icon, List } from "@raycast/api";
+import { ActionPanel, Action, Icon, List, getPreferenceValues } from "@raycast/api";
 
 import { User, useChannels } from "./shared/client";
 import { withSlackClient } from "./shared/withSlackClient";
 import { useFrecencySorting } from "@raycast/utils";
 import { OpenChannelInSlack, OpenChatInSlack, useSlackApp } from "./shared/OpenInSlack";
+import { convertSlackEmojiToUnicode, getTimeLocale } from "./shared/utils";
+
+const { displayExtraMetadata } = getPreferenceValues();
+
+function getCoworkerTime(coworkerTimeZone: string): string {
+  const time = new Date();
+  return new Intl.DateTimeFormat(getTimeLocale(), {
+    timeZone: coworkerTimeZone,
+    hour: "2-digit",
+    minute: "2-digit",
+  }).format(time);
+}
+
+function searchItemAccessories(
+  statusEmoji: string | undefined,
+  statusText: string | undefined,
+  statusExpiration: string | null,
+  timezone: string,
+) {
+  const searchMetadata: Array<{ icon: string | Icon; text: string; tooltip?: string | null | undefined }> = [
+    {
+      icon: convertSlackEmojiToUnicode(statusEmoji ?? ""),
+      text: statusText ?? "",
+      tooltip: statusExpiration ? String(statusExpiration) : undefined,
+    },
+  ];
+
+  if (displayExtraMetadata) {
+    searchMetadata.push(
+      { icon: Icon.Globe, text: timezone.split("/")[1].replace(/_/g, " ") },
+      { icon: Icon.Clock, text: getCoworkerTime(timezone) },
+    );
+  }
+
+  return searchMetadata;
+}
 
 function Search() {
   const { isAppInstalled, isLoading } = useSlackApp();
@@ -21,12 +57,25 @@ function Search() {
         const isUser = item.id.startsWith("U");
 
         if (isUser) {
-          const { id: userId, name, icon, teamId: workspaceId, conversationId } = item as User;
+          const {
+            id: userId,
+            name,
+            icon,
+            title,
+            statusEmoji,
+            statusText,
+            statusExpiration,
+            teamId: workspaceId,
+            conversationId,
+            timezone,
+          } = item as User;
           return (
             <List.Item
               key={userId}
               title={name}
+              subtitle={displayExtraMetadata ? title : undefined}
               icon={icon}
+              accessories={searchItemAccessories(statusEmoji, statusText, statusExpiration, timezone)}
               actions={
                 <ActionPanel>
                   <OpenChatInSlack