+++ extensions/slack/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Slack Changelog
 
+## [Update to Search command] - {PR_MERGE_DATE}  
+- Adds job title, timezone, current time, and slack status to the List that gets returned as part of the Search command.
+- Adds preferences to show this metadata, as well as time format preferences throughout the extension. 
+- Adds Emojify to unread messages command to convert :fireworks: to 🎆, for example.
+
 ## [Search Messages command] - 2024-07-08
 
 - Add a new command to search through your Slack's workspace messages.